Separate tabs into files & fix calibration not working properly when facing other directions
